Move 404 redirect logic to getServerSidePropsWrapper function
